pyproject.toml
[project]
name = "lrgwn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "hydra-core>=1.3.2",
    "scipy>=1.17.0",
    "torch>=2.10.0",
    "torch-geometric>=2.7.0",
]
